Extract theme-settings from settings-view
