+++ src/assets/main.css
@@ -67,7 +67,13 @@ p {
   margin: 5px 0;
 }
 
+/*Fix time picker style*/
 .v-time-picker-clock,
 .v-time-picker-controls__time__field .v-field {
   background-color: rgba(var(--v-theme-primary), 0.1);
+}
+
+/* Don't use harsh white for v-switch label*/
+.v-selection-control .v-label {
+  color: rgb(var(--v-theme-on-surface-variant))
 }
